Fix duplicate dependency issue
We mis-analyzed the previous error. The error is still the same.

profile::swh::deploy::storage and profile::swh::deploy::storage_archiver
shares the same dependency on python3-swh.storage package.

Puppet won't accept to solve it.  As a fix, we remove such dependency in
profile::swh::deploy::storage_archiver since we already have it with
profile::swh::deploy::storage.

Note that it is not ideal.  The better approach would be to package the
archiver on its own package and then refer to such dependency in the
manifest.

Related T481