data/defaults: Evenly share index workers amongst azure nodes
The orchestrator (all contents) instanciates the indexer referenced in
its setup file (here the mimetype one).  As a result, there need to
exist those indexers' setup files in that node.  Otherwise, it just
fails (worker07 and worker08 in the setup prior to this commit for
example).

This is not enforced in the manifest for now.

As a result, the install is like this:
- mimetype indexer + indexer orchestrator alongside (odd node number)
- language indexer + indexer orchestrator (text) alongside (even node
  number)