pgbouncer: Run pgbouncer as postgres user
First things first, iso production deployment.

At the moment, the pgbouncer module wants (reasonably) a dedicated `pgbouncer`
user to run it (and that fails as this user does not exist in our manifest).

I'm currently modifying this default setup to run as current production
runs (`postgres` user runs it).

This can be adapted later if everything is fine and that needs to be changed.

Related T1784