Allow overriding the order of interface blocks in profile::network
There is a race condition between vlan and bond interfaces when using ifupdown
on buster. Unfortunately, the default orders on the trepasi-debnet modules are
backwards (they put static ifaces before manual ones) so the network won't start
if you have a manual iface on top of a vlan iface on top of a bond.

Hack the network profile to counteract this default ordering, and hardcode the
order of interface blocks for the kafka hosts which have such a layering of
interfaces.