Puppet manage elasticsearch nodes
This allows declaration of the elasticsearch instances which was mostly
configured manually so far. We reused the actual production configuration in
/etc/elasticsearch/{elasticsearch.yml,jvm_options} as default for the
production nodes.

For both the production and staging nodes, the following commit configures:
- /etc/elasticsearch/elasticsearch.yml (overriding the one from the debian
  package)
- /etc/elasticsearch/jvm.options.d/jvm.options

This also fixed a couple of current paper cuts:
- fix the inter-dependency on package/service/apt-config order
- remove a deprecated xpack configuration (since 7.8.0 which is the prod
  version)
- unmanage the no longer required openjdk-8 dependency

Manual actions to undertake when this diff lands:
- uninstall the openjdk-8 from the production esnodes
- apply the following configuration in production one node at a time.

Related to T2817