backfill: Dedicate storage configuration for the backfiller
Prior to this commit, we would share the storage configuration for both storage and the
backfiller. This would end up having possibly a journal_writer configuration key on some
instances.

This reworks the configuration to allow a read-only storage configuration. And installs
the required logging configuration file.

Related to T3009