Install update-metrics as a service called daily
As an implementation detail, the service is called daily through a systemd timer.

This updates the scheduler_metrics table regularly so the distribution per forge stays
up-to-date. This also avoid having this running in a tmux session on saatchi. This also
deals with properly update such metrics on the staging infra.

Related to T3127