Install save code now update routine as a service called every minute
As an implementation detail, the service is called every minute through a systemd timer.
This means that if a timer is already running after 1 minute, this won't start
immediately. This will avoid overlapping updates altogether.

As another pros, as this becomes a systemd service, this will keep the running logs.

Related to T3451