Empty unneeded env variable so celery does not complain about it
Prior to this commit, a plain empty variable was not enough and celery when started
would complain about it and not start.

This actually renames and moves the variable at the end of the cli call. It makes it
"empty" as "--" when not needed. That way celery actually agrees to start in both
cases (used or unused).

This also renames it to something not bound to celery to avoid potential celery extra
behavior on particularly named env variables.