Unstuck opam loader loading for the default opam.ocaml.org instance
This is a workaround that needs reverting when opam loader (and opam) will be able to
properly deal with extraction of metadata for a package (from another opam instance than
the default one).

Currently the loader does a computation to determine the path of the package. This is ok
for other instance but not the default one. So this creates a symbolic link from the
"default" instance folder to another named after the "opam.ocaml.org" so everything is
ok.

Related to T3717