sanoid: Add the configuration to manage the snapshot retention on backup01
Not happy with the way the configuration is done because it forces
to declare part of configuration on the 2 servers.
It's error prone and to know exactly the backup workflow to add a zfs
backup policy.
It will definitely need some improvements but it will only worth it
if we start to use this kind of backup massively.

Related to T3889