Bootstrap elastic worker role
Those worker nodes should be able to run docker containers. As an implementation detail,
we mount the main docker partition as zfs volume with compression enable to keep disk
use in check. It's expected they have an extra /dev/vdb partition to be
provided (through the provisioner vagrant or terraform-proxmox).

Related to T4144