staging: Activate metadata loader for git loader
This will install the extra python3-swh.loader.metadata package for staging workers.

By default, this is off for production and on for staging.

When on, it's also installing the extra set of credentials in the loader_git's
configuration to allow the use of such credentials when communicating with the forges.

Related to T4193