recurrent: Adapt default policy so origins without last update are scheduled
The current default policy is not appropriate no origins without last update are
scheduled at all.

We have some listers [1] which are not able to provide the last update for some origins.
We should reschedule those once in a while so they can get ingested too.

[1] conda, bower, cran, ...

Refs. swh/infra/sysadm-environment#4971