Install webapp icinga checks as static checks
As we are migrating to the elastic infra, we cannot rely on puppet resource collection
anymore.

That's pulled by the actual migration of the staging webapp. The current production
webapp (moma as archive.s.o) is still running on the static infra. As this will need to
change as soon as we migrate it (soonish) to the elastic infra, it's simpler for the
puppet manifest to migrate now.

Refs. swh/infra/sysadm-environment#5095