Allow webapp icinga checks to work
Currently, icinga refuses to start because it does not know the staging webapp Host
object (and other dependencies). It's indeed no longer puppet managed so it's not
installed any long.

This also adapts the icinga checks for the archive.s.o to work (by mentioning where it
runs, moma). And currently, with this small patch, icinga no longer complains about it.
It was complaining the same way as for webapp.staging without it.

Refs. swh/infra/sysadm-environment#5095