staging: Make db1 a rancher node
We'll keep running the staging dbs on the node as usual (no rancher just yet).

The goal is to add an objstorage on that node, to relieve storage1's objstorage which is
almost full. We'll keep storage1's objstorage as a read-only instance. And create a
multiplexer objstorage so workers can still read/write on those composed objstorage
instances.

Refs. swh/infra/sysadm-environment#5260